--- sds-extension/sds-apache-dubbo-2.7.5/src/main/resources/META-INF/dubbo/org.apache.dubbo.rpc.Filter
@@ -1,4 +1,0 @@
-SdsProviderFilter=com.didiglobal.sds.extension.dubbo.filter.SdsProviderFilter
-SdsConsumerFilter=com.didiglobal.sds.extension.dubbo.filter.SdsConsumerFilter
-
-
