--- sds-extension/sds-apache-dubbo-common/src/main/java/com/didiglobal/sds/extension/dubbo/constant/SdsDubboConstants.java
@@ -1,19 +1,0 @@
-package com.didiglobal.sds.extension.dubbo.constant;
-
-/**
- * Created by sea on 2020-03-28.
- */
-public class SdsDubboConstants {
-
-    public static final int SDS_PROVIDER_FILTER_ORDER = -10100;
-
-    public static final int SDS_CONSUMER_FILTER_ORDER = -10050;
-
-    public static final String SDS_DUBBO_POINT = "sdsDubboPoint";
-
-    public static final String SDS_DOWNGRADE_START_TIME = "sdsDowngradeStartTime";
-
-    private SdsDubboConstants() {
-
-    }
-}
