+++ sds-extension/sds-apache-dubbo/src/main/resources/META-INF/dubbo/org.apache.dubbo.rpc.Filter
@@ -1,0 +1,3 @@
+SdsApacheDubboFilter=com.didiglobal.sds.extension.dubbo.filter.SdsApacheDubboFilter
+
+
