+++ sds-extension/sds-dubbo/src/main/resources/META-INF/dubbo/com.alibaba.dubbo.rpc.Filter
@@ -1,4 +1,3 @@
-SdsProviderFilter=com.didiglobal.sds.extension.dubbo.filter.SdsProviderFilter
-SdsConsumerFilter=com.didiglobal.sds.extension.dubbo.filter.SdsConsumerFilter
+SdsDubboFilter=com.didiglobal.sds.extension.dubbo.filter.SdsDubboFilter
